# frontend/internal/api/openapi.snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Label Printer Hub \u2014 backend",
    "description": "REST + SSE API for the Label Printer Hub backend. The Go frontend consumes the OpenAPI spec at /openapi.json via oapi-codegen; humans browse the interactive docs at /docs (Swagger UI) or /redoc.",
    "version": "0.0.0.dev0"
  },
  "paths": {
...
    "/print": {
      "post": {
        "tags": [
          "print"
        ],
        "summary": "Submit a print job",
        "description": "Submit a label-print job.  The job is queued and dispatched asynchronously by the queue worker.  Returns 202 with the new job's UUID and state ``queued``.  Returns 4xx/5xx on printer errors (tape mismatch, offline, cover open, etc.).",
        "operationId": "create_print_job_print_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrintRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrintJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
    "/api/printers": {
      "get": {
        "tags": [
          "printers"
        ],
        "summary": "List all printers",
        "description": "Returns every registered printer.  The ``paused`` flag is joined from ``printer_state``; it is ``false`` when no state row exists yet.  Pass ``?slug=<slug>`` to filter to a single printer by exact slug match (returns 404 when no printer with that slug exists).",
        "operationId": "list_printers_api_printers_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true,
              "description": "Filter by exact slug",
              "title": "Slug"
            },
            "description": "Filter by exact slug"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/app__schemas__printer__PrinterRead"
                  },
                  "title": "Response List Printers Api Printers Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/printers/{printer_id}": {
      "get": {
        "tags": [
          "printers"
        ],
        "summary": "Get printer detail",
        "description": "Returns full metadata for a single printer, including the ``paused`` flag joined from ``printer_state``. Returns 404 when the printer is not registered.",
        "operationId": "get_printer_api_printers__printer_id__get",
        "parameters": [
          {
            "name": "printer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Printer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__schemas__printer__PrinterRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/printers": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Alle Drucker auflisten",
        "description": "Gibt alle Drucker zur\u00fcck. Deaktivierte Drucker werden standardm\u00e4\u00dfig ausgeblendet. Mit ``?include_disabled=true`` werden auch deaktivierte Drucker zur\u00fcckgegeben.",
        "operationId": "list_printers_api_v1_admin_printers_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "include_disabled",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Disabled"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/app__api__routes__admin_printers_api__PrinterRead"
                  },
                  "title": "Response List Printers Api V1 Admin Printers Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Neuen Drucker anlegen",
        "description": "Legt einen neuen Drucker an. Slug und Name m\u00fcssen eindeutig sein. Gibt 409 zur\u00fcck wenn Slug oder Name bereits vergeben ist.",
        "operationId": "create_printer_api_v1_admin_printers_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrinterCreatePayload"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__api__routes__admin_printers_api__PrinterRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/printers/{slug}": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Einzelnen Drucker abrufen",
        "description": "Gibt einen Drucker per Slug zur\u00fcck. 404 wenn kein Drucker mit diesem Slug existiert.",
        "operationId": "get_printer_api_v1_admin_printers__slug__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__api__routes__admin_printers_api__PrinterRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "Drucker aktualisieren",
        "description": "Aktualisiert einen Drucker per PATCH-Semantik (nur ge\u00e4nderte Felder). Slug und ID k\u00f6nnen nicht ge\u00e4ndert werden. Gibt 404 zur\u00fcck wenn kein Drucker mit diesem Slug existiert.",
        "operationId": "update_printer_api_v1_admin_printers__slug__put",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrinterUpdatePayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__api__routes__admin_printers_api__PrinterRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/printers/{slug}/disable": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Drucker deaktivieren",
        "description": "Deaktiviert einen Drucker (Soft-Delete). Gibt 404 zur\u00fcck wenn kein Drucker mit diesem Slug existiert. Gibt 409 zur\u00fcck wenn der Drucker bereits deaktiviert ist.",
        "operationId": "disable_printer_api_v1_admin_printers__slug__disable_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__api__routes__admin_printers_api__PrinterRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/printers/{slug}/enable": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Drucker aktivieren",
        "description": "Aktiviert einen deaktivierten Drucker. Gibt 404 zur\u00fcck wenn kein Drucker mit diesem Slug existiert. Gibt 409 zur\u00fcck wenn der Drucker bereits aktiv ist.",
        "operationId": "enable_printer_api_v1_admin_printers__slug__enable_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__api__routes__admin_printers_api__PrinterRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ContentType": {
        "type": "string",
        "enum": [
          "qr_only",
          "qr_one_line",
          "qr_two_lines",
          "qr_three_lines",
          "text_one_line",
          "text_two_lines",
          "qr_with_listing"
        ],
        "title": "ContentType",
        "description": "Tape-independent semantic content types for label rendering."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PrintJobResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "const": "queued",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status"
        ],
        "title": "PrintJobResponse",
        "description": "POST /print 202 body \u2014 queue accepted."
      },
...
      "PrintLookupRequest": {
        "properties": {
          "app": {
            "type": "string",
            "title": "App"
          },
          "identifier": {
            "type": "string",
            "title": "Identifier"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "app",
          "identifier"
        ],
        "title": "PrintLookupRequest",
        "description": "Resolve label data via an integration plugin."
      },
      "PrintOptions": {
        "properties": {
          "copies": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Copies",
            "default": 1
          },
          "auto_cut": {
            "type": "boolean",
            "title": "Auto Cut",
            "default": true
          },
          "high_resolution": {
            "type": "boolean",
            "title": "High Resolution",
            "default": false
          },
          "half_cut": {
            "type": "boolean",
            "title": "Half Cut",
            "default": false
          },
          "last_page": {
            "type": "boolean",
            "title": "Last Page",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "PrintOptions",
        "description": "Per-print options \u2014 copies, cut behaviour, resolution."
      },
      "PrintRequest": {
        "properties": {
          "content_type": {
            "$ref": "#/components/schemas/ContentType"
          },
          "options": {
            "$ref": "#/components/schemas/PrintOptions",
            "default": {
              "copies": 1,
              "auto_cut": true,
              "high_resolution": false,
              "half_cut": false,
              "last_page": true
            }
          },
          "data": {
            "$ref": "#/components/schemas/RawLabelData",
            "nullable": true
          },
          "lookup": {
            "$ref": "#/components/schemas/PrintLookupRequest",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "content_type"
        ],
        "title": "PrintRequest",
        "description": "POST /api/print body.\n\nEither `data` (RawLabelData) or `lookup` (PrintLookupRequest) is provided.\nExactly one of the two must be present."
      },
      "PrinterConnection": {
        "properties": {
          "host": {
            "type": "string",
            "maxLength": 253,
            "minLength": 1,
            "title": "Host"
          },
          "port": {
            "type": "integer",
            "maximum": 65535.0,
            "minimum": 1.0,
            "title": "Port"
          },
          "snmp": {
            "$ref": "#/components/schemas/SNMPConfig"
          }
        },
        "type": "object",
        "required": [
          "host",
          "port"
        ],
        "title": "PrinterConnection",
        "description": "Verbindungsparameter f\u00fcr einen Drucker."
      },
      "PrinterCreatePayload": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "pattern": "^[a-z0-9][a-z0-9-]{1,62}[a-z0-9]$",
            "title": "Slug"
          },
          "model": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Model"
          },
          "backend": {
            "type": "string",
            "enum": [
              "ptouch",
              "brother_ql"
            ],
            "title": "Backend"
          },
          "connection": {
            "$ref": "#/components/schemas/PrinterConnection"
          },
          "queue": {
            "$ref": "#/components/schemas/PrinterQueueSettings"
          },
          "cut_defaults": {
            "$ref": "#/components/schemas/PrinterCutDefaults"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "name",
          "slug",
          "model",
          "backend",
          "connection"
        ],
        "title": "PrinterCreatePayload",
        "description": "Payload f\u00fcr das Anlegen eines neuen Druckers via Admin-API."
      },
...
      "PrinterQueueSettings": {
        "properties": {
          "timeout_s": {
            "type": "integer",
            "maximum": 600.0,
            "minimum": 1.0,
            "title": "Timeout S",
            "default": 30
          }
        },
        "type": "object",
        "title": "PrinterQueueSettings",
        "description": "Warteschlangen-Einstellungen f\u00fcr einen Drucker."
      },
...
      "PrinterUpdatePayload": {
        "properties": {
          "name": {
            "type": "string",
            "nullable": true,
            "title": "Name"
          },
          "connection": {
            "$ref": "#/components/schemas/PrinterConnection",
            "nullable": true
          },
          "queue": {
            "$ref": "#/components/schemas/PrinterQueueSettings",
            "nullable": true
          },
          "cut_defaults": {
            "$ref": "#/components/schemas/PrinterCutDefaults",
            "nullable": true
          },
          "enabled": {
            "type": "boolean",
            "nullable": true,
            "title": "Enabled"
          }
        },
        "type": "object",
        "title": "PrinterUpdatePayload",
        "description": "Payload f\u00fcr das Aktualisieren eines bestehenden Druckers via Admin-API.\n\nDer Service ignoriert stillschweigend: slug, model, backend, id.\nAlle Felder sind optional \u2014 ein leerer Body ist ein g\u00fcltiger PATCH."
      },
...
      "RawLabelData": {
        "properties": {
          "title": {
            "type": "string",
            "nullable": true,
            "title": "Title"
          },
          "primary_id": {
            "type": "string",
            "nullable": true,
            "title": "Primary Id"
          },
          "qr_payload": {
            "type": "string",
            "nullable": true,
            "title": "Qr Payload"
          },
          "secondary": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Secondary",
            "default": []
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/LabelDataItem"
            },
            "type": "array",
            "title": "Items",
            "default": []
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "RawLabelData",
        "description": "Raw label payload accepted when the client supplies data directly.\n\nMirrors LabelData minus `source_app` (set server-side to 'manual').\nAll content fields are optional \u2014 ContentType-specific validation\nhappens in LayoutEngine._validate_data."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "app__api__routes__admin_printers_api__PrinterRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "backend": {
            "type": "string",
            "title": "Backend"
          },
          "connection": {
            "additionalProperties": true,
            "type": "object",
            "title": "Connection"
          },
          "queue": {
            "additionalProperties": true,
            "type": "object",
            "title": "Queue"
          },
          "cut_defaults": {
            "additionalProperties": true,
            "type": "object",
            "title": "Cut Defaults"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "slug",
          "model",
          "backend",
          "connection",
          "queue",
          "cut_defaults",
          "enabled",
          "created_at",
          "updated_at"
        ],
        "title": "PrinterRead",
        "description": "Lesbare Darstellung eines Druckers.\n\nEnth\u00e4lt alle DB-Felder \u2014 keine internen Implementierungsdetails."
      },
      "app__schemas__printer__PrinterRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug",
            "default": ""
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "backend": {
            "type": "string",
            "title": "Backend"
          },
          "connection": {
            "additionalProperties": true,
            "type": "object",
            "title": "Connection"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "paused": {
            "type": "boolean",
            "title": "Paused"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "model",
          "backend",
          "connection",
          "enabled",
          "paused",
          "created_at",
          "updated_at"
        ],
        "title": "PrinterRead",
        "description": "Full representation of a Printer row, augmented with the paused flag.\n\n``paused`` is joined from the ``printer_state`` table; it defaults to\n``False`` for printers whose state row was not yet created (safe \u2014 the\nDB lifespan helper creates state rows at startup, so this only matters\nin tests or during the very first boot)."
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Label-Hub-Key"
      }
    }
  }
}